+++ py_autotask/__init__.py
@@ -26,7 +26,7 @@ Example:
     })
 """
 
-__version__ = "2.1.0"
+__version__ = "1.0.0"
 __author__ = "Aaron Sachs"
 __email__ = "asachs@wyre.engineering"